Genrative Models/GAN_classifier/requirements.txt
numpy==1.16.5
opencv-python==4.2.0.32
keras==2.3.1
tensorflow-gpu==1.15
matplotlib==3.1.1
scikit-learn==0.21.3